backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.1.0
drf-base64==2.0
Pillow==9.4.0
webcolors==1.11.1